requirements.dev.txt
bandit~=1.7.5
black~=23.11.0
boto3-stubs~=1.29.6
bpython~=0.24
flake8~=6.1.0
isort~=5.12.0
mypy~=1.7.0
pycodestyle~=2.11.1
pylint~=3.0.2
pytest~=7.4.3
